# backend/swagger_documentation.yaml
openapi: 3.0.1
info:
  title: WaterFilterBusiness.API
  version: '1.0'
paths:
...
  /api/CustomerCalls/customers/{customerId}:
    get:
      tags:
        - CustomerCalls
      parameters:
        - name: customerId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
            default: 1
      responses:
        '200':
          description: Success
  /api/Customers:
    get:
      tags:
        - Customers
      parameters:
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
            default: 1
        - name: excludeWithCalls
          in: query
          style: form
          schema:
            type: boolean
            default: true
        - name: excludeRedListed
          in: query
          style: form
          schema:
            type: boolean
            default: true
        - name: includeOnlyRedListed
          in: query
          style: form
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Success
    post:
      tags:
        - Customers
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
          text/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
          application/*+json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
      responses:
        '200':
          description: Success
  /api/Customers/{id}:
    get:
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
    patch:
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequestModel'
      responses:
        '200':
          description: Success
  /api/Customers/{id}/change-history:
    get:
      tags:
        - Customers
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
...
  /api/InventoryItems:
    post:
      tags:
        - InventoryItems
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem_AddRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/InventoryItem_AddRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/InventoryItem_AddRequestModel'
      responses:
        '200':
          description: Success
    get:
      tags:
        - InventoryItems
      parameters:
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
  /api/InventoryItems/{id}:
    delete:
      tags:
        - InventoryItems
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
    patch:
      tags:
        - InventoryItems
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem_PatchRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/InventoryItem_PatchRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/InventoryItem_PatchRequestModel'
      responses:
        '200':
          description: Success
...
  /api/InventoryPurchases:
    get:
      tags:
        - InventoryPurchases
      parameters:
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
    post:
      tags:
        - InventoryPurchases
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryPurchase_AddRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/InventoryPurchase_AddRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/InventoryPurchase_AddRequestModel'
      responses:
        '200':
          description: Success
...
  /api/ScheduledCalls/phone-operator/{phoneOperatorId}:
    get:
      tags:
        - ScheduledCalls
      parameters:
        - name: phoneOperatorId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: paginationCursor
          in: query
          style: form
          schema:
            type: integer
            format: int32
            default: 0
        - name: scheduledFor
          in: query
          style: form
          schema:
            type: string
            format: date
            example: 0001-1-1
      responses:
        '200':
          description: Success
  /api/ScheduledCalls:
    post:
      tags:
        - ScheduledCalls
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledCall_AddRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/ScheduledCall_AddRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ScheduledCall_AddRequestModel'
      responses:
        '200':
          description: Success
  /api/ScheduledCalls/{id}:
    delete:
      tags:
        - ScheduledCalls
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
...
  /api/TechnicianInventoryItems/technician/{technicianId}:
    get:
      tags:
        - TechnicianInventoryItems
      parameters:
        - name: technicianId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
  /api/TechnicianInventoryItems/technician/{technicianId}/tool/{toolId}/reduce-stock:
    patch:
      tags:
        - TechnicianInventoryItems
      parameters:
        - name: technicianId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
        - name: toolId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int32
          text/json:
            schema:
              type: integer
              format: int32
          application/*+json:
            schema:
              type: integer
              format: int32
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int32
        fullName:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        profession:
          type: string
          nullable: true
        isQualified:
          type: boolean
          nullable: true
        redListedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    CustomerUpdateRequestModel:
      type: object
      properties:
        fullName:
          maxLength: 55
          minLength: 0
          type: string
          nullable: true
        phoneNumber:
          maxLength: 30
          minLength: 0
          type: string
          nullable: true
        address:
          maxLength: 70
          minLength: 0
          type: string
          nullable: true
        city:
          maxLength: 70
          minLength: 0
          type: string
          nullable: true
        profession:
          maxLength: 65
          minLength: 0
          type: string
          nullable: true
        isQualified:
          type: boolean
          nullable: true
        redListedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    InventoryItem_AddRequestModel:
      required:
        - name
        - price
      type: object
      properties:
        name:
          maxLength: 45
          minLength: 0
          type: string
        price:
          minimum: 0.0001
          type: number
          format: double
      additionalProperties: false
    InventoryItem_PatchRequestModel:
      type: object
      properties:
        name:
          maxLength: 45
          minLength: 0
          type: string
          nullable: true
        price:
          minimum: 0.0001
          type: number
          format: double
          nullable: true
      additionalProperties: false
    InventoryPurchase_AddRequestModel:
      required:
        - toolId
      type: object
      properties:
        toolId:
          type: integer
          format: int32
      additionalProperties: false
...
    ScheduledCall_AddRequestModel:
      required:
        - customerId
        - phoneAgentId
        - scheduledAt
      type: object
      properties:
        customerId:
          type: integer
          format: int32
        phoneAgentId:
          type: integer
          format: int32
        scheduledAt:
          type: string
          format: date-time
      additionalProperties: false
...